+++ spec/rubocop/cop/lint/parentheses_as_grouped_expression_spec.rb
@@ -74,6 +74,17 @@ RSpec.describe RuboCop::Cop::Lint::ParenthesesAsGroupedExpression do
     expect_no_offenses('assert_equal (0..1.9), acceleration.domain')
   end
 
+  it 'does not register an offesne when heredoc has a space between the same string as the method name and `(`' do
+    expect_no_offenses(<<~RUBY)
+      foo(
+        <<~EOS
+          foo (
+          )
+        EOS
+      )
+    RUBY
+  end
+
   context 'when using safe navigation operator' do
     it 'registers an offense and corrects for method call with space before ' \
        'the parenthesis' do